+++ samplesheets/assayapps/microarray/plugins.py
@@ -103,13 +103,14 @@ class SampleSheetAssayPlugin(SampleSheetAssayPluginPoint):
 
         return None
 
-    def update_row(self, row, table, assay):
+    def update_row(self, row, table, assay, index):
         """
         Update render table row with e.g. links. Return the modified row.
 
         :param row: Original row (list of dicts)
-        :param table: Full table (dict)
+        :param table: Full table (list of lists)
         :param assay: Assay object
+        :param index: Row index (int)
         :return: List of dicts
         """
         assay_path = self.get_assay_path(assay)