+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Changed
     - Remove Bootstrap tooltip updating for iRODS buttons (#1949)
 - **Samplesheets**
     - Sanitize iRODS paths in ``get_row_path()`` calls (#1947)
+    - ``index`` arg in ``SampleSheetAssayPluginPoint.update_row()`` (#1957)
 - **Taskflowbackend**
     - Increase default for ``TASKFLOW_IRODS_CONN_TIMEOUT`` (#1900)
     - Disable lock requirement for role and project update flows (#1948)